moved apiRoutertest into a test folder
