+++ package.json
@@ -21,6 +21,7 @@
     "bcrypt": "^5.0.1",
     "concurrently": "^7.0.0",
     "cors": "^2.8.5",
+    "dotenv": "^16.0.0",
     "express": "^4.17.3",
     "jest": "^24.9.0",
     "jsonwebtoken": "^8.5.1",
